durin_configurator/requirements.txt
pyserial
pycapnp